requirements.txt
python-binance
mysql-connector-python
python-telegram-bot